backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0
langdetect>=1.0.9
python-dotenv>=1.0.1
newsapi-python>=0.2.7
beautifulsoup4>=4.12.3
